Add switch --dry-run

+++ dotfilesmanager/ioutils.py
@@ -19,8 +19,10 @@ def eprint(*args, **kwargs):
 
 
 def _create_file(file_name, contents):
-    with open(file_name, 'w') as file:
-        file.write(contents)
+    sprint("Writing to file: {0}".format(file_name))
+    if not env.ARGS.dry_run:
+        with open(file_name, 'w') as file:
+            file.write(contents)
 
 
 def _back_up_file(file_name):
@@ -30,10 +32,13 @@ def _back_up_file(file_name):
         file_name[file_name.rfind('/') + 1:]
         .replace('.', '') + '_' +
         timestamp + '.bak')
-    sprint("\tBacking up " + file_name + " to " + bak_file)
     if not os.path.exists(env.BACKUPS_DIR):
-        os.mkdir(env.BACKUPS_DIR)
-    shutil.move(file_name, bak_file)
+        sprint("Creating backups dir {0}".format(env.BACKUPS_DIR))
+        if not env.ARGS.dry_run:
+            os.mkdir(env.BACKUPS_DIR)
+    sprint("\tBacking up {0} to {1}".format(file_name, bak_file))
+    if not env.ARGS.dry_run:
+        shutil.move(file_name, bak_file)
 
 
 def compile_dotfile(file_name, input_files):
@@ -63,9 +68,10 @@ def _write_input_file_contents(file_name, out_buffer):
 def _write_output_file(file_path, contents):
     if not env.ARGS.clobber and isfile(file_path):
         _back_up_file(file_path)
-    sprint("\tWriting output file " + file_path)
-    with open(file_path, 'w') as output_file:
-        output_file.write(contents.getvalue())
+    sprint("\tWriting input file contents to output file " + file_path)
+    if not env.ARGS.dry_run:
+        with open(file_path, 'w') as output_file:
+            output_file.write(contents.getvalue())
 
 
 def _write_to_output_buffer(output, file_buffer):
@@ -82,9 +88,13 @@ def revert_dotfile(dotfile):
         while choice not in (['Y', 'N']):
             choice = input(
                 "Revert {0} to backup located at {1}? (Y/N): "
-                .format(dotfile, bak_file)).upper()
+                    .format(dotfile, bak_file)).upper()
             if choice == 'Y':
-                os.remove(join(env.OUTPUT_DIR, dotfile))
-                shutil.copy(bak_file, join(env.OUTPUT_DIR, dotfile))
+                existing_dotfile = join(env.OUTPUT_DIR, dotfile)
+                sprint("Removing dotfile {0} and replacing with backup named {1}"
+                        .format(existing_dotfile, bak_file))
+                if not env.ARGS.dry_run:
+                    os.remove(existing_dotfile)
+                    shutil.copy(bak_file, join(env.OUTPUT_DIR, dotfile))
     else:
         eprint("No backup files found matching {0}".format(dotfile))
